+++ src/Middleware.test.ts
@@ -1,0 +1,5 @@
+import { Middleware } from './Middleware';
+
+test('There is an error if you instantiate Middleware class', () => {
+	expect(() => new Middleware()).toThrowError();
+});