+++ src/controller/ControllerBase.test.ts
@@ -1,0 +1,5 @@
+import { ControllerBase } from './ControllerBase';
+
+test('if exception will be thrown for direct ControllerBase instantiating', () => {
+	expect(() => new ControllerBase({} as any)).toThrowError();
+});